2024.11.3 로또 결과 출력 기능 완료
