+++ src/main/java/lotto/controller/LottoController.java
@@ -1,4 +1,11 @@
-package lotto;
+package lotto.controller;
+
+import lotto.dao.LottoDAO;
+import lotto.service.LottoCreateService;
+import lotto.service.LottoMatchingSerivce;
+import lotto.service.LottoPurchaseService;
+import lotto.service.LottoSelectService;
+import lotto.view.LottoView;
 
 public class LottoController {
         private final LottoView lottoView = new LottoView();
@@ -6,6 +13,7 @@ public class LottoController {
         private final LottoPurchaseService lottoPurchaseService = new LottoPurchaseService();
         private final LottoCreateService lottoCreateService=new LottoCreateService();
         private final LottoSelectService lottoSelectService=new LottoSelectService();
+        private final LottoMatchingSerivce lottoMatchingSerivce= new LottoMatchingSerivce();
 
         public void lottoMain() {
                 try {
@@ -16,7 +24,9 @@ public class LottoController {
                    lottoView.printLotto(lottoDAO);
                    lottoDAO.setLottoSelect(lottoSelectService.getLottoSelect(lottoView.getLottoSelected()));
                    lottoDAO.setBonusNumber(lottoView.getLottoBonus());
-
+                   lottoDAO.setLottoResult(lottoMatchingSerivce.getLottoResult(lottoDAO.getLottoDatabase(),lottoDAO.getLottoSelect(),lottoDAO.getBonusNumber()));
+                   lottoDAO.setRateReturn(lottoMatchingSerivce.getRateReturn(lottoDAO.getLottoResult(),lottoDAO.getCost()));
+                   lottoView.printLottoResult(lottoDAO.getLottoResult(),lottoDAO.getRateReturn());
 
                 } catch (IllegalArgumentException e) {
                         System.out.println(e.getMessage());
