+++ src/main/java/lotto/vo/Lotto.java
@@ -1,5 +1,6 @@
-package lotto;
+package lotto.vo;
 
+import java.util.HashSet;
 import java.util.List;
 
 public class Lotto {
@@ -13,6 +14,14 @@ public class Lotto {
     private void validate(List<Integer> numbers) {
         if (numbers.size() != 6) {
             throw new IllegalArgumentException("[ERROR] 로또 번호는 6개여야 합니다.");
+        }
+        if (numbers.size() != new HashSet<>(numbers).size()) {
+            throw new IllegalArgumentException("[ERROR] 로또 번호는 중복될 수 없습니다.");
+        }
+        for (int number : numbers) {
+            if (number < 1 || number > 45) {
+                throw new IllegalArgumentException("[ERROR] 로또 번호는 1부터 45 사이여야 합니다.");
+            }
         }
     }
